Channel Data Requests - YouTube Data API

This is a function to make API calls to the YouTube Data API for channel specific data. While it can be called directly, it is primarily intended to help get lists of videos within a channel. For more in-depth use of the Data API, please refer to tuber. For more details on the following arguments, please refer to https://developers.google.com/youtube/v3/docs/channels/list

Usage

data_channel_request(
  part = NULL,
  categoryId = NULL,
  forUsername = NULL,
  hl = NULL,
  id = NULL,
  managedByMe = NULL,
  maxResults = 5,
  mine = NULL,
  onBehalfOfContentOwner = NULL,
  pageToken = NULL,
  token = getOption("YouTube_Token")
)

Arguments

part

Required. String.

categoryId

String.

forUsername

String.

hl

String.

id

String.

managedByMe

String.

maxResults

Integer.

mine

String.

onBehalfOfContentOwner

String.

pageToken

String.

token

Required.

Value

data.frame

Examples

if (FALSE) { # \dontrun{
data_channel_request(part = "contentDetails",
                     mine = "true")
} # }